fix #35
