+++ tests/index.jsx
@@ -5,7 +5,14 @@ console.warn = function (warning) {
   warn.apply(console, arguments);
 };
 
-require('./CreateForm.test.jsx');
 require('./convertSchema.test');
 require('./ErrorMessage.test.jsx');
 require('./Validator.test');
+require('./CreateForm.test.jsx');
+require('./didSubmit.test.jsx');
+require('./linkField.test.jsx');
+require('./submitForm.test.jsx');
+require('./submitGroup.test.jsx');
+require('./validateField.test.jsx');
+require('./isGroupValid.test.jsx');
+require('./isValid.test.jsx');
