+++ tests/isGroupValid.test.jsx
@@ -1,0 +1,14 @@
+var should = require('should');
+var React = require('react/addons');
+var TestUtils = React.addons.TestUtils;
+var assign = require('react/lib/Object.assign');
+
+describe('CreateForm', function () {
+
+  var Formation = require('../src/form');
+
+  describe('#isGroupValid', function () {
+    //TODO
+  });
+
+});
